+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Changed
+- Reverted the optimizing phase to being invoked by Apply on Play when VRCFury is not installed (#135)
 
 ### Removed